remove deep nested folders and components inside members feature
