+++ src/features/members/Members.js
@@ -1,7 +1,7 @@
 import Logo from 'layout/Logo';
 import PageTitle from 'layout/PageTitle';
 import PageContainer from 'layout/PageContainer';
-import MembersMainContainer from './mainContainer/MembersMainContainer';
+import MembersMainContainer from './MembersMainContainer';
 
 const Members = () => {
   return (
