--- src/features/members/mainContainer/actionSection/index.js
@@ -1,1 +1,0 @@
-export { default } from './ActionSection';
